loading/lazyload/attribute_off_iframe.tentative.sub.html

<!DOCTYPE html>
<head>
  <title>Loading iframes with lazyload="off"</title>
  <link rel="author" title="Scott Little" href="mailto:sclittle@chromium.org">
  <link rel="help" href="https://github.com/whatwg/html/pull/3752">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<body>
  <iframe id="in_viewport" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_1.html" width="200px" height="100px" lazyload="off"></iframe>
  <div style="height:10000px;"></div>
  <iframe id="below_viewport" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_2.html" width="200px" height="100px" lazyload="off"></iframe>
</body>

<script>
  var t = async_test("Test that both iframes load immediately and block the window's load event.");

  var has_in_viewport_iframe_loaded = false;
  document.getElementById("in_viewport").addEventListener("load",
    t.step_func(function() {
      assert_false(has_in_viewport_iframe_loaded);
      has_in_viewport_iframe_loaded = true;
    }));

  var has_below_viewport_iframe_loaded = false;
  document.getElementById("below_viewport").addEventListener("load",
    t.step_func(function() {
      assert_false(has_below_viewport_iframe_loaded);
      has_below_viewport_iframe_loaded = true;
    }));

  window.addEventListener("load", t.step_func_done(function() {
    assert_true(has_in_viewport_iframe_loaded);
    assert_true(has_below_viewport_iframe_loaded);
  }));
</script>
